ARM: OMAP: calculate hsmmc data timeout accurately
The data timeout can be calculated from the clock
rate and timing values given by the card.  At
present the timeout is set to the maximum value,
about 2.7 secs.  When dealing with cards that do have
errors, this causes large unnecessary delays because
the correct timeout may be around 10 millisecs.

Signed-off-by: Adrian Hunter <ext-adrian.hunter@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>