bq27x00_battery: Check current after reading flags
It appears bq27500 has a threshold for DSG flag (60mA by default according
to the manual) after which it no longer sets DSG bit. This means we have
to check current to avoid misreporting that we are charging when we are
really discharging, which currently happens in lower power states.
As flags register is read in bq27x00_update, do current direction check
there.