ARM: OMAP2+: PM: Remove bogus fiq_[enable/disable] tuple
On OMAP platform, FIQ is reserved for secure environment only. If at all
the FIQ needs to be disabled, it involves going through security
API call. Hence the local_fiq_[enable/disable]() in the OMAP code is bogus.
On GP devices too, the fiq is disabled for non-secure software.

So just get rid of it.

[notasas@gmail.com: 3.2 backport]
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>

Conflicts:

	arch/arm/mach-omap2/cpuidle34xx.c
	arch/arm/mach-omap2/cpuidle44xx.c
	arch/arm/mach-omap2/pm24xx.c
	arch/arm/mach-omap2/pm34xx.c
	arch/arm/mach-omap2/pm44xx.c