wl1251: implement disconnect command properly
Currently the driver uses a disconnect command with 4 data bytes
by reusing wl1251_cmd_template_set(), however the data length is
supposed to be 8 bytes. This doesn't seem to causing any bad effects,
but it's better to send correct commands.

Tested on pandora board to confirm that idle mode is still working
after this patch with no change in current draw.