ARM: dma-mapping: avoid speculative prefetching fix
This partially reverts 2ffe2da3e71652d4f4cae19539b5c78c2a239136.

Cortex-A8 doesn't seem to be doing any speculative data prefetching,
there is no mention of this in the manual (unlike as for A9), only
instruction prefetch is mentioned. For us this fix is only causing
useless performance penalty, avoid it. The 2.6.27 kernel we used before
didn't have this fix and all seemed to be fine.